requirements.txt
django==4.1.3
djangorestframework
requests